requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pathlib2>=2.3.7
typing-extensions>=4.5.0